+++ home/scripts/gentoo.sh
@@ -2,42 +2,45 @@
 
 set -o errexit -o nounset
 
+. $HOME/.local/share/chezmoi/home/scripts/lib.sh
+
 ins="emerge -av --changed-use"
 pkgs=""
 USE_DIR="/etc/portage/package.use"
-
-. $HOME/.local/share/chezmoi/home/scripts/lib.sh
+AUTH=$(search_auth)
 
 euse_pkg() {
   if ! grep -q "$2" "$USE_DIR"/"${1#*/}" 2>/dev/null ; then
-    sudo euse -p "$1" -E "$2"
+    "$AUTH" euse -p "$1" -E "$2"
   fi
 }
 
 euse_global() {
   if ! grep -q "$1" /etc/portage/make.conf ; then
-    sudo euse -E "$1"
+    "$AUTH" euse -E "$1"
   fi
 }
 
 install_deps() {
   # env
-  if ! hash euse 2>/dev/null; then sudo $ins gentoolkit ; fi
+  if ! hash euse 2>/dev/null; then "$AUTH" $ins gentoolkit ; fi
   euse_global jpeg
 
-  [ -d "$USE_DIR" ] || sudo mkdir -p "$USE_DIR"
-  [ -d /etc/portage/package.accept_keyworks ] || sudo mkdir -p /etc/portage/package.accept_keywords
+  [ -d "$USE_DIR" ] || "$AUTH" mkdir -p "$USE_DIR"
+  [ -d /etc/portage/package.accept_keyworks ] || "$AUTH" mkdir -p /etc/portage/package.accept_keywords
 
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.accept_keywords/* /etc/portage/package.accept_keywords/
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/dotfiles "$USE_DIR/"
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.accept_keywords/* /etc/portage/package.accept_keywords/
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/dotfiles "$USE_DIR/"
 
   if has_systemd ; then
-    sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-systemd "$USE_DIR/"
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-systemd "$USE_DIR/"
+    pkgs="brave-bin"
   else
-    sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-elogind "$USE_DIR/"
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-elogind "$USE_DIR/"
+    pkgs"librewolf"
   fi
 
-  pkgs="app-crypt/gnupg pass zsh awesome media-sound/mpd ncmpcpp xinit xorg-server xst
+  pkgs="$pkgs app-crypt/gnupg pass zsh awesome media-sound/mpd ncmpcpp xinit xorg-server xst
     feh picom maim vifm mpv zathura zathura-pdf-mupdf
     neomutt cava ueberzug weechat net-misc/yt-dlp
     papirus-icon-theme media-sound/mpc lightdm inotify-tools light stow
@@ -46,27 +49,26 @@ install_deps() {
 }
 
 install_pulse() {
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/pulseaudio "$USE_DIR/"
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/pulseaudio "$USE_DIR/"
 
-  pkgs="$pkgs pulseaudio firefox-bin"
+  pkgs="$pkgs pulseaudio"
 }
 
 install_alsa() {
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/alsa "$USE_DIR/"
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/alsa "$USE_DIR/"
 
-  pkgs="$pkgs alsa-utils tap-plugins swh-plugins libsamplerate cmt-plugins caps-plugins ladspa-bs2b alsa-plugins"
-  # need to be update from ninjatool >  brave-bin
+  pkgs="$pkgs alsa-utils tap-plugins swh-plugins libsamplerate cmt-plugins ladspa-bs2b alsa-plugins"
 }
 
 install_emacs() {
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/emacs "$USE_DIR/"
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/emacs "$USE_DIR/"
 
   # discount = markdown
   pkgs="$pkgs ripgrep discount emacs"
 }
 
 install_vim() {
-  sudo cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/vim "$USE_DIR/"
+  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/vim "$USE_DIR/"
 
   pkgs="$pkgs vim"
 }
@@ -116,7 +118,7 @@ main() {
   "$EMACS" && install_emacs
   "$EXTRA" && install_extra_deps
 
-  sudo $ins $pkgs
+  "$AUTH" $ins $pkgs
 
   exit 0
 }